+++ scripts/bootstrap.sh
@@ -14,8 +14,12 @@ if [[ ! -d Common.Mod.Example/run/Mods ]]; then
 fi
 
 # Vintage Story files
-cp "${VINTAGE_STORY}/Lib/protobuf-net.dll" vendor/protobuf-net.dll
-cp "${VINTAGE_STORY}/VintagestoryAPI.dll" vendor/VintagestoryAPI.dll
+if [[ ! -f vendor/protobuf-net.dll ]]; then
+  cp "${VINTAGE_STORY}/Lib/protobuf-net.dll" vendor/protobuf-net.dll
+fi
+if [[ ! -f vendor/VintagestoryAPI.dll ]]; then
+  cp "${VINTAGE_STORY}/VintagestoryAPI.dll" vendor/VintagestoryAPI.dll
+fi
 
 # 3rd party mods
 ## ImGui
